splitting reference_context into three modules
