+++ isovar/allele_reads.py
@@ -34,121 +34,114 @@ from .string_helpers import convert_from_bytes_if_necessary, trim_N_nucleotides
 
 logger = logging.getLogger(__name__)
 
-
 # subclassing from namedtuple to get a lightweight object with built-in
 # hashing and comparison while also being able to add methods
-AlleleReadFields = namedtuple(
-    "AlleleRead",
-    "prefix allele suffix name sequence")
+class AlleleRead(
+        namedtuple("AlleleRead", "prefix allele suffix name sequence")):
 
-class AlleleRead(AlleleReadFields):
-    def __new__(cls, prefix, allele, suffix, name):
-        # construct sequence from prefix + alt + suffix
-        return AlleleReadFields.__new__(
-            cls,
-            prefix=prefix,
-            allele=allele,
-            suffix=suffix,
-            name=name,
-            sequence=prefix + allele + suffix)
+    @property
+    def sequence(self):
+        return self.prefix + self.allele + self.suffix
 
     def __len__(self):
         return len(self.prefix) + len(self.allele) + len(self.suffix)
 
-def allele_read_from_locus_read(locus_read, n_ref):
-    """
-    Given a single ReadAtLocus object, return either an AlleleRead or None
+    @classmethod
+    def from_locus_read(cls, locus_read, n_ref):
+        """
+        Given a single LocusRead object, return either an AlleleRead or None
 
-    Parameters
-    ----------
-    locus_read : LocusRead
-        Read which overlaps a variant locus but doesn't necessarily contain the
-        alternate nucleotides
+        Parameters
+        ----------
+        locus_read : LocusRead
+            Read which overlaps a variant locus but doesn't necessarily contain the
+            alternate nucleotides
 
-    n_ref : int
-        Number of reference positions we are expecting to be modified or
-        deleted (for insertions this should be 0)
-    """
-    sequence = locus_read.sequence
-    reference_positions = locus_read.reference_positions
+        n_ref : int
+            Number of reference positions we are expecting to be modified or
+            deleted (for insertions this should be 0)
+        """
+        sequence = locus_read.sequence
+        reference_positions = locus_read.reference_positions
 
-    # positions of the nucleotides before and after the variant within
-    # the read sequence
-    read_pos_before = locus_read.base0_read_position_before_variant
-    read_pos_after = locus_read.base0_read_position_after_variant
+        # positions of the nucleotides before and after the variant within
+        # the read sequence
+        read_pos_before = locus_read.base0_read_position_before_variant
+        read_pos_after = locus_read.base0_read_position_after_variant
 
-    # positions of the nucleotides before and after the variant on the
-    # reference genome
-    ref_pos_before = reference_positions[read_pos_before]
+        # positions of the nucleotides before and after the variant on the
+        # reference genome
+        ref_pos_before = reference_positions[read_pos_before]
 
-    if ref_pos_before is None:
-        logger.warn(
-            "Missing reference pos for nucleotide before variant on read: %s",
-            locus_read)
-        return None
-
-    ref_pos_after = reference_positions[read_pos_after]
-
-    if ref_pos_after is None:
-        logger.warn(
-            "Missing reference pos for nucleotide after variant on read: %s",
-            locus_read)
-        return None
-
-    if n_ref == 0:
-        if ref_pos_after - ref_pos_before != 1:
-            # if the number of nucleotides skipped isn't the same
-            # as the number of reference nucleotides in the variant then
-            # don't use this read
-            logger.debug(
-                "Positions before (%d) and after (%d) variant should be adjacent on read %s",
-                ref_pos_before,
-                ref_pos_after,
+        if ref_pos_before is None:
+            logger.warn(
+                "Missing reference pos for nucleotide before variant on read: %s",
                 locus_read)
             return None
 
-        # insertions require a sequence of non-aligned bases
-        # followed by the subsequence reference position
-        ref_positions_for_inserted = reference_positions[
-            read_pos_before + 1:read_pos_after]
-        if any(insert_pos is not None for insert_pos in ref_positions_for_inserted):
-            # all these inserted nucleotides should *not* align to the
-            # reference
-            logger.debug(
-                "Skipping read, inserted nucleotides shouldn't map to reference")
-            return None
-    else:
-        # substitutions and deletions
-        if ref_pos_after - ref_pos_before != n_ref + 1:
-            # if the number of nucleotides skipped isn't the same
-            # as the number of reference nucleotides in the variant then
-            # don't use this read
-            logger.debug(
-                "Positions before (%d) and after (%d) variant should be adjacent on read %s",
-                ref_pos_before,
-                ref_pos_after,
+        ref_pos_after = reference_positions[read_pos_after]
+
+        if ref_pos_after is None:
+            logger.warn(
+                "Missing reference pos for nucleotide after variant on read: %s",
                 locus_read)
             return None
 
-    nucleotides_at_variant_locus = sequence[read_pos_before + 1:read_pos_after]
+        if n_ref == 0:
+            if ref_pos_after - ref_pos_before != 1:
+                # if the number of nucleotides skipped isn't the same
+                # as the number of reference nucleotides in the variant then
+                # don't use this read
+                logger.debug(
+                    "Positions before (%d) and after (%d) variant should be adjacent on read %s",
+                    ref_pos_before,
+                    ref_pos_after,
+                    locus_read)
+                return None
 
-    prefix = sequence[:read_pos_before + 1]
-    suffix = sequence[read_pos_after:]
+            # insertions require a sequence of non-aligned bases
+            # followed by the subsequence reference position
+            ref_positions_for_inserted = reference_positions[
+                read_pos_before + 1:read_pos_after]
+            if any(insert_pos is not None for insert_pos in ref_positions_for_inserted):
+                # all these inserted nucleotides should *not* align to the
+                # reference
+                logger.debug(
+                    "Skipping read, inserted nucleotides shouldn't map to reference")
+                return None
+        else:
+            # substitutions and deletions
+            if ref_pos_after - ref_pos_before != n_ref + 1:
+                # if the number of nucleotides skipped isn't the same
+                # as the number of reference nucleotides in the variant then
+                # don't use this read
+                logger.debug(
+                    ("Positions before (%d) and after (%d) variant should be "
+                     "adjacent on read %s"),
+                    ref_pos_before,
+                    ref_pos_after,
+                    locus_read)
+                return None
 
-    prefix, suffix = convert_from_bytes_if_necessary(prefix, suffix)
-    prefix, suffix = trim_N_nucleotides(prefix, suffix)
+        nucleotides_at_variant_locus = sequence[read_pos_before + 1:read_pos_after]
 
-    return AlleleRead(
-        prefix,
-        nucleotides_at_variant_locus,
-        suffix,
-        name=locus_read.name)
+        prefix = sequence[:read_pos_before + 1]
+        suffix = sequence[read_pos_after:]
+
+        prefix, suffix = convert_from_bytes_if_necessary(prefix, suffix)
+        prefix, suffix = trim_N_nucleotides(prefix, suffix)
+
+        return cls(
+            prefix,
+            nucleotides_at_variant_locus,
+            suffix,
+            name=locus_read.name)
 
 def allele_reads_from_locus_reads(locus_reads, n_ref):
     """
-    Given a collection of ReadAtLocus objects, returns a
-    list of VariantRead objects (which are split into prefix/variant/suffix
-    nucleotides).
+    Given a collection of LocusRead objects, returns a
+    list of AlleleRead objects
+    (which are split into prefix/allele/suffix nucleotide strings).
 
     Parameters
     ----------
@@ -161,7 +154,7 @@ def allele_reads_from_locus_reads(locus_reads, n_ref):
     """
 
     for locus_read in locus_reads:
-        allele_read = allele_read_from_locus_read(locus_read, n_ref)
+        allele_read = AlleleRead.from_locus_read(locus_read, n_ref)
         if allele_read is None:
             continue
         else:
