+++ isovar/locus_reads.py
@@ -33,154 +33,154 @@ from .dataframe_builder import DataFrameBuilder
 
 logger = logging.getLogger(__name__)
 
-
-LocusRead = namedtuple(
-    "LocusRead",
-    [
+class LocusRead(namedtuple("LocusRead", [
         "name",
         "sequence",
         "reference_positions",
         "quality_scores",
         "base0_read_position_before_variant",
-        "base0_read_position_after_variant",
-    ])
+        "base0_read_position_after_variant"])):
 
-def create_locus_read_from_pysam_pileup_element(
-        pileup_element,
-        base0_position_before_variant,
-        base0_position_after_variant,
-        use_secondary_alignments,
-        use_duplicate_reads,
-        min_mapping_quality):
-    """
-    Parameters
-    ----------
-    pileup_element : pysam.PileupRead
+    @classmethod
+    def from_pysam_pileup_element(
+            cls,
+            pileup_element,
+            base0_position_before_variant,
+            base0_position_after_variant,
+            use_secondary_alignments,
+            use_duplicate_reads,
+            min_mapping_quality):
+        """
+        Parameters
+        ----------
+        pileup_element : pysam.PileupRead
 
-    base0_position_before_variant : int
+        base0_position_before_variant : int
 
-    base0_position_after_variant : int
+        base0_position_after_variant : int
 
-    use_secondary_alignments : bool
+        use_secondary_alignments : bool
 
-    use_duplicate_reads : bool
+        use_duplicate_reads : bool
 
-    min_mapping_quality : int
+        min_mapping_quality : int
 
-    Returns LocusRead or None
-    """
-    read = pileup_element.alignment
+        Returns LocusRead or None
+        """
+        read = pileup_element.alignment
 
-    # For future reference,  may get overlapping reads
-    # which can be identified by having the same name
-    name = read.query_name
+        # For future reference,  may get overlapping reads
+        # which can be identified by having the same name
+        name = read.query_name
 
-    if name is None:
-        logger.warn("Read missing name at position %d",
-            base0_position_before_variant + 1)
-        return None
+        if name is None:
+            logger.warn(
+                "Read missing name at position %d",
+                base0_position_before_variant + 1)
+            return None
 
-    if read.is_unmapped:
-        logger.warn(
-            "How did we get unmapped read '%s' in a pileup?", name)
-        return None
+        if read.is_unmapped:
+            logger.warn(
+                "How did we get unmapped read '%s' in a pileup?", name)
+            return None
 
-    if pileup_element.is_refskip:
-        # if read sequence doesn't actually align to the reference
-        # base before a variant, skip it
-        logger.debug("Skipping pileup element with CIGAR alignment N (intron)")
-        return None
-    elif pileup_element.is_del:
-        logger.debug(
-            "Skipping deletion at position %d (read name = %s)",
+        if pileup_element.is_refskip:
+            # if read sequence doesn't actually align to the reference
+            # base before a variant, skip it
+            logger.debug("Skipping pileup element with CIGAR alignment N (intron)")
+            return None
+        elif pileup_element.is_del:
+            logger.debug(
+                "Skipping deletion at position %d (read name = %s)",
                 base0_position_before_variant + 1,
                 name)
-        return None
+            return None
 
-    if read.is_secondary and not use_secondary_alignments:
-        logger.debug("Skipping secondary alignment of read '%s'", name)
-        return None
+        if read.is_secondary and not use_secondary_alignments:
+            logger.debug("Skipping secondary alignment of read '%s'", name)
+            return None
 
-    if read.is_duplicate and not use_duplicate_reads:
-        logger.debug("Skipping duplicate read '%s'", name)
-        return None
+        if read.is_duplicate and not use_duplicate_reads:
+            logger.debug("Skipping duplicate read '%s'", name)
+            return None
 
-    mapping_quality = read.mapping_quality
+        mapping_quality = read.mapping_quality
 
-    missing_mapping_quality = mapping_quality is None
-    if min_mapping_quality > 0 and missing_mapping_quality:
-        logger.debug("Skipping read '%s' due to missing MAPQ", name)
-        return None
-    elif mapping_quality < min_mapping_quality:
-        logger.debug(
-            "Skipping read '%s' due to low MAPQ: %d < %d",
+        missing_mapping_quality = mapping_quality is None
+        if min_mapping_quality > 0 and missing_mapping_quality:
+            logger.debug("Skipping read '%s' due to missing MAPQ", name)
+            return None
+        elif mapping_quality < min_mapping_quality:
+            logger.debug(
+                "Skipping read '%s' due to low MAPQ: %d < %d",
                 read.mapping_quality,
                 mapping_quality,
                 min_mapping_quality)
-        return None
+            return None
 
-    sequence = read.query_sequence
+        sequence = read.query_sequence
 
-    if sequence is None:
-        logger.warn("Read '%s' missing sequence", name)
-        return None
+        if sequence is None:
+            logger.warn("Read '%s' missing sequence", name)
+            return None
 
-    base_qualities = read.query_qualities
+        base_qualities = read.query_qualities
 
-    if base_qualities is None:
-        logger.warn("Read '%s' missing base qualities", name)
-        return None
+        if base_qualities is None:
+            logger.warn("Read '%s' missing base qualities", name)
+            return None
 
-    #
-    # Documentation for pysam.AlignedSegment.get_reference_positions:
-    # ------------------------------------------------------------------
-    # By default, this method only returns positions in the reference
-    # that are within the alignment. If full_length is set, None values
-    # will be included for any soft-clipped or unaligned positions
-    # within the read. The returned list will thus be of the same length
-    # as the read.
-    #
-    # Source:
-    # http://pysam.readthedocs.org/en/latest/
-    # api.html#pysam.AlignedSegment.get_reference_positions
-    #
-    # We want a None value for every read position that does not have a
-    # corresponding reference position.
-    reference_positions = read.get_reference_positions(
-        full_length=True)
+        #
+        # Documentation for pysam.AlignedSegment.get_reference_positions:
+        # ------------------------------------------------------------------
+        # By default, this method only returns positions in the reference
+        # that are within the alignment. If full_length is set, None values
+        # will be included for any soft-clipped or unaligned positions
+        # within the read. The returned list will thus be of the same length
+        # as the read.
+        #
+        # Source:
+        # http://pysam.readthedocs.org/en/latest/
+        # api.html#pysam.AlignedSegment.get_reference_positions
+        #
+        # We want a None value for every read position that does not have a
+        # corresponding reference position.
+        reference_positions = read.get_reference_positions(
+            full_length=True)
 
-    # pysam uses base-0 positions everywhere except region strings
-    # Source:
-    # http://pysam.readthedocs.org/en/latest/faq.html#pysam-coordinates-are-wrong
-    if base0_position_before_variant not in reference_positions:
-        logger.debug(
-            "Skipping read '%s' because first position %d not mapped",
+        # pysam uses base-0 positions everywhere except region strings
+        # Source:
+        # http://pysam.readthedocs.org/en/latest/faq.html#pysam-coordinates-are-wrong
+        if base0_position_before_variant not in reference_positions:
+            logger.debug(
+                "Skipping read '%s' because first position %d not mapped",
                 name,
                 base0_position_before_variant)
-        return None
-    else:
-        base0_read_position_before_variant = reference_positions.index(
-            base0_position_before_variant)
+            return None
+        else:
+            base0_read_position_before_variant = reference_positions.index(
+                base0_position_before_variant)
 
-    if base0_position_after_variant not in reference_positions:
-        logger.debug(
-            "Skipping read '%s' because last position %d not mapped",
+        if base0_position_after_variant not in reference_positions:
+            logger.debug(
+                "Skipping read '%s' because last position %d not mapped",
                 name,
                 base0_position_after_variant)
-        return None
-    else:
-        base0_read_position_after_variant = reference_positions.index(
-            base0_position_after_variant)
+            return None
+        else:
+            base0_read_position_after_variant = reference_positions.index(
+                base0_position_after_variant)
 
-    if isinstance(sequence, bytes):
-        sequence = sequence.decode('ascii')
-    return LocusRead(
-        name=name,
-        sequence=sequence,
-        reference_positions=reference_positions,
-        quality_scores=base_qualities,
-        base0_read_position_before_variant=base0_read_position_before_variant,
-        base0_read_position_after_variant=base0_read_position_after_variant)
+        if isinstance(sequence, bytes):
+            sequence = sequence.decode('ascii')
+
+        return cls(
+            name=name,
+            sequence=sequence,
+            reference_positions=reference_positions,
+            quality_scores=base_qualities,
+            base0_read_position_before_variant=base0_read_position_before_variant,
+            base0_read_position_after_variant=base0_read_position_after_variant)
 
 def pileup_reads_at_position(samfile, chromosome, base0_position):
     """
@@ -250,9 +250,9 @@ def locus_read_generator(
 
     logger.debug(
         "Gathering reads at locus %s: %d-%d",
-            chromosome,
-            base1_position_before_variant,
-            base1_position_after_variant)
+        chromosome,
+        base1_position_before_variant,
+        base1_position_after_variant)
     base0_position_before_variant = base1_position_before_variant - 1
     base0_position_after_variant = base1_position_after_variant - 1
 
@@ -267,7 +267,7 @@ def locus_read_generator(
             samfile=samfile,
             chromosome=chromosome,
             base0_position=base0_position_before_variant):
-        read = create_locus_read_from_pysam_pileup_element(
+        read = LocusRead.from_pysam_pileup_element(
             pileup_element,
             base0_position_before_variant=base0_position_before_variant,
             base0_position_after_variant=base0_position_after_variant,
@@ -281,10 +281,10 @@ def locus_read_generator(
 
     logger.info(
         "Found %d reads overlapping locus %s: %d-%d",
-            count,
-            chromosome,
-            base1_position_before_variant,
-            base1_position_after_variant)
+        count,
+        chromosome,
+        base1_position_before_variant,
+        base1_position_after_variant)
 
 def locus_reads_dataframe(*args, **kwargs):
     """
