+++ isovar/reference_context.py
@@ -16,64 +16,12 @@ from __future__ import print_function, division, absolute_import
 from collections import namedtuple, OrderedDict, defaultdict
 import logging
 
-from skbio import DNA
-
 from .effect_prediction import reference_transcripts_for_variant
-from .variant_helpers import interbase_range_affected_by_variant_on_transcript
 from .dataframe_builder import DataFrameBuilder
-
+from .reference_sequence_key_with_reading_frame import (
+    ReferenceSequenceKeyWithReadingFrame,)
 
 logger = logging.getLogger(__name__)
-
-
-##########################
-#
-# SequenceKey
-# -----------
-#
-# Used to identify and group the distinct sequences occurring  on a set of
-# transcripts overlapping a variant locus
-#
-##########################
-
-SequenceKey = namedtuple(
-    "SequenceKey", [
-        "strand",
-        "sequence_before_variant_locus",
-        "sequence_at_variant_locus",
-        "sequence_after_variant_locus"
-    ]
-)
-
-
-##########################
-#
-# SequenceKeyWithReadingFrame
-# ---------------------------
-#
-# Includes all the fields of a SequenceKey, but also includes a reading frame
-# at the start of the reference sequence.
-#
-#
-##########################
-
-SequenceKeyWithReadingFrame = namedtuple(
-    "ReferenceContext", SequenceKey._fields + (
-        # if the reference context includes the 5' UTR then
-        # this is the offset to the start codon, otherwise it's the
-        # offset needed to get the first base of a codon
-        "offset_to_first_complete_codon",
-        # does this context overlap a start codon?
-        "overlaps_start_codon",
-        # does this context contain the whole trinucleotide start codon?
-        "contains_start_codon",
-        # does this context contain any UTR bases?
-        "contains_five_prime_utr",
-        # translation of complete codons in the reference context
-        # before the variant
-        "amino_acids_before_variant",
-    )
-)
 
 
 ##########################
@@ -89,249 +37,10 @@ SequenceKeyWithReadingFrame = namedtuple(
 
 ReferenceContext = namedtuple(
     "ReferenceContext",
-    SequenceKeyWithReadingFrame._fields + (
+    ReferenceSequenceKeyWithReadingFrame._fields + (
         "variant",
         "transcripts")
 )
-
-def variant_matches_reference_sequence(variant, ref_seq_on_transcript, strand):
-    """
-    Make sure that reference nucleotides we expect to see on the reference
-    transcript from a variant are the same ones we encounter.
-    """
-    if strand == "-":
-        ref_seq_on_transcript = str(
-            DNA(ref_seq_on_transcript).reverse_complement())
-
-    return ref_seq_on_transcript == variant.ref
-
-def sequence_key_for_variant_on_transcript(variant, transcript, context_size):
-    """
-    Extracts the reference sequence around a variant locus on a particular
-    transcript.
-
-    Parameters
-    ----------
-    variant : varcode.Variant
-
-    transcript : pyensembl.Transcript
-
-    context_size : int
-
-    Returns SequenceKey object with the following fields:
-        - strand
-        - sequence_before_variant_locus
-        - sequence_at_variant_locus
-        - sequence_after_variant_locus
-
-    Can also return None if Transcript lacks sufficiently long sequence
-    """
-
-    full_sequence = transcript.sequence
-
-    if full_sequence is None:
-        logger.warn(
-            "Expected transcript %s (overlapping %s) to have sequence",
-                transcript,
-                variant)
-        return None
-
-    if len(full_sequence) < 6:
-        # need at least 6 nucleotides for a start and stop codon
-        logger.warn(
-            "Sequence of %s (overlapping %s) too short: %d",
-                transcript,
-                variant,
-                len(full_sequence))
-        return None
-
-    # get the interbase range of offsets which capture all reference
-    # bases modified by the variant
-    variant_start_offset, variant_end_offset = \
-        interbase_range_affected_by_variant_on_transcript(
-            variant=variant,
-            transcript=transcript)
-
-    logger.info("Interbase offset range on %s for variant %s = %d:%d",
-        transcript,
-        variant,
-        variant_start_offset,
-        variant_end_offset)
-
-    prefix = full_sequence[
-        max(0, variant_start_offset - context_size):
-        variant_start_offset]
-
-    suffix = full_sequence[
-        variant_end_offset:
-        variant_end_offset + context_size]
-
-    ref_nucleotides_at_variant = full_sequence[
-        variant_start_offset:variant_end_offset]
-    if not variant_matches_reference_sequence(
-            variant=variant,
-            strand=transcript.strand,
-            ref_seq_on_transcript=ref_nucleotides_at_variant):
-        # TODO: once we're more confident about other logic in isovar,
-        # change this to a warning and return None to allow for modest
-        # differences between reference sequence patches, since
-        # GRCh38.p1 may differ at some positions from GRCh38.p5
-        raise ValueError(
-            "Wrong reference sequence for variant %s on transcript %s" % (
-                variant,
-                transcript))
-    return SequenceKey(
-        strand=transcript.strand,
-        sequence_before_variant_locus=prefix,
-        sequence_at_variant_locus=ref_nucleotides_at_variant,
-        sequence_after_variant_locus=suffix)
-
-
-def reading_frame_to_offset(reading_frame_at_start_of_sequence):
-    """
-    Given a reading frame (how many nucleotides into a codon) at the
-    start of a cDNA sequence, return the number of nucleotides which need
-    to be trimmed to start on a complete codon.
-
-    Parameters
-    ----------
-
-    reading_frame_at_start_of_sequence : int
-
-    Returns an int
-    """
-    if reading_frame_at_start_of_sequence < 0:
-        raise ValueError("Reading frame can't be negative: %d" % (
-            reading_frame_at_start_of_sequence,))
-    elif reading_frame_at_start_of_sequence > 2:
-        raise ValueError("Reading frame must be within 0 and 2, not %d" % (
-            reading_frame_at_start_of_sequence,))
-    # If we're 1 nucleotide into the codon then we need to shift
-    # over two more to restore the ORF. Likewise, if we're 2 nucleotides in
-    # then we have to shift over one more.
-    return (3 - reading_frame_at_start_of_sequence) % 3
-
-
-def sequence_key_with_reading_frame_for_variant_on_transcript(
-        variant,
-        transcript,
-        context_size):
-    """
-    Extracts the reference sequence around a variant locus on a particular
-    transcript and determines the reading frame at the start of that
-    sequence context.
-
-    Parameters
-    ----------
-    variant : varcode.Variant
-
-    transcript : pyensembl.Transcript
-
-    context_size : int
-
-    Returns SequenceKeyWithReadingFrame object or None if Transcript lacks
-    coding sequence, protein sequence or annotated start/stop codons.
-    """
-    if not transcript.contains_start_codon:
-        logger.info(
-            "Expected transcript %s for variant %s to have start codon",
-                transcript,
-                variant)
-        return None
-
-    if not transcript.contains_stop_codon:
-        logger.info(
-            "Expected transcript %s for variant %s to have stop codon",
-                transcript,
-                variant)
-        return None
-
-    if not transcript.protein_sequence:
-        logger.info(
-            "Expected transript %s for variant %s to have protein sequence",
-                transcript,
-                variant)
-        return None
-
-    sequence_key = sequence_key_for_variant_on_transcript(
-        variant=variant,
-        transcript=transcript,
-        context_size=context_size)
-
-    if sequence_key is None:
-        logger.info("No sequence key for variant %s on transcript %s",
-            variant,
-            transcript)
-        return None
-
-    # get the interbase range of offsets which capture all reference
-    # bases modified by the variant
-    variant_start_offset, variant_end_offset = \
-        interbase_range_affected_by_variant_on_transcript(
-            variant=variant,
-            transcript=transcript)
-
-    start_codon_idx = min(transcript.start_codon_spliced_offsets)
-
-    # skip any variants which occur in the 5' UTR or overlap the start codon
-    # since
-    #   (1) UTR variants have unclear coding effects and
-    #   (2) start-loss variants may result in a new start codon / reading
-    #       frame but we're not sure which one!
-    if variant_start_offset < start_codon_idx + 3:
-        logger.info(
-            "Skipping transcript %s for variant %s, must be after start codon",
-                transcript,
-                variant)
-        return None
-
-    stop_codon_idx = min(transcript.stop_codon_spliced_offsets)
-
-    # skip variants which affect the 3' UTR of the transcript since
-    # they don't have obvious coding effects on the protein sequence
-    if variant_start_offset >= stop_codon_idx + 3:
-        logger.info(
-            "Skipping transcript %s for variant %s, occurs in 3' UTR",
-                transcript,
-                variant)
-        return None
-
-    n_prefix = len(sequence_key.sequence_before_variant_locus)
-    prefix_start_idx = variant_start_offset - n_prefix
-    n_bases_between_start_and_variant = variant_start_offset - start_codon_idx
-    n_full_codons_before_variant = n_bases_between_start_and_variant // 3
-
-    # if the sequence before the variant contains more bases than the
-    # distance to the start codon, then by definition it must contain
-    # some untranslated bases
-    contains_five_prime_utr = (n_prefix > n_bases_between_start_and_variant)
-    # allows for the possibility that the first base in the sequence might
-    # be the first nucleotide of the start codon
-    contains_start_codon = (n_prefix >= n_bases_between_start_and_variant)
-    # the sequence context might only include the 2nd or 3rd bases of
-    # the start codon
-    overlaps_start_codon = (n_prefix > n_bases_between_start_and_variant - 3)
-
-    if contains_start_codon:
-        offset_to_first_complete_codon = start_codon_idx - prefix_start_idx
-        amino_acids_before_variant = transcript.protein_sequence[:n_full_codons_before_variant]
-    else:
-        reading_frame = (prefix_start_idx - start_codon_idx) % 3
-        offset_to_first_complete_codon = reading_frame_to_offset(reading_frame)
-        n_codons_in_prefix = (n_prefix - offset_to_first_complete_codon) // 3
-        amino_acids_before_variant = transcript.protein_sequence[
-            n_full_codons_before_variant - n_codons_in_prefix:
-            n_full_codons_before_variant]
-    return SequenceKeyWithReadingFrame(
-        strand=sequence_key.strand,
-        sequence_before_variant_locus=sequence_key.sequence_before_variant_locus,
-        sequence_at_variant_locus=sequence_key.sequence_at_variant_locus,
-        sequence_after_variant_locus=sequence_key.sequence_after_variant_locus,
-        offset_to_first_complete_codon=offset_to_first_complete_codon,
-        contains_start_codon=contains_start_codon,
-        overlaps_start_codon=overlaps_start_codon,
-        contains_five_prime_utr=contains_five_prime_utr,
-        amino_acids_before_variant=amino_acids_before_variant)
 
 def sort_key_decreasing_max_length_transcript_cds(reference_context):
     """
@@ -367,7 +76,7 @@ def reference_contexts_for_variant(
 
     for transcript in overlapping_transcripts:
         sequence_key_with_reading_frame = \
-            sequence_key_with_reading_frame_for_variant_on_transcript(
+            ReferenceSequenceKeyWithReadingFrame.from_variant_and_transcript(
                 variant=variant,
                 transcript=transcript,
                 context_size=context_size)
